reports/requirements.txt
requests
ujson
aws-psycopg2
dnspython
marshmallow
pyjwt
pytest
